chore(audit): structure audit 2026-06-09 + C6/C7 cheap-sweep
Fresh codebase-wide structural pass over the areas the 2026-06-08 audit
did not cover (src/compute/ 40k LOC, the loader family, header hygiene,
tests). Findings in docs/audit/structural_debt_2026_06_09.md; several
agent-reported "dead flags/files" were false positives (search omitted
tools/ or referenced an already-deleted tree) and are recorded under
Refuted so they are not re-chased.

Opening cheap sweep (verbatim, behaviour-neutral):

C6 — exec/inference_state.h forward-declares JsonConstrainer /
SchemaConstrainer instead of including their ~190-line headers; those
rode transitively into every executor.h consumer. executor.cu gains a
direct compute/schema_constrain.h include (it dereferences
schema_constrainer->apply_mask); engine_scheduler.cpp only assigns and
null-checks the pointers, so the forward decl suffices.

C7 — new lightweight tests/test_cuda_skip.h owns the single
HasCudaDevice() + SKIP_IF_NO_CUDA(); removed 11 local macro definitions
(three divergent bodies) + 6 local HasCudaDevice() re-definitions. Kept
separate from the heavy test_model_builder.h so host-compiled .cpp tests
don't pull in cuda_fp16/half just to gate on a device.

make build (CUDA 13.3) + make test-unit (37/37) green.
