+++ tests/test_continuous_batching.cpp
@@ -6,6 +6,7 @@
 #include "runtime/scheduler.h"
 #include "memory/kv_cache.h"
 #include "memory/kv_cache_manager.h"
+#include "test_cuda_skip.h"
 
 #include <cstdint>
 #include <memory>
@@ -13,22 +14,6 @@
 
 namespace imp {
 namespace {
-
-// ============================================================================
-// Helper: skip the entire test if no CUDA device is available.
-// ============================================================================
-static bool HasCudaDevice() {
-    int count = 0;
-    cudaError_t err = cudaGetDeviceCount(&count);
-    return err == cudaSuccess && count > 0;
-}
-
-#define SKIP_IF_NO_CUDA()                               \
-    do {                                                \
-        if (!HasCudaDevice()) {                         \
-            GTEST_SKIP() << "No CUDA device available"; \
-        }                                               \
-    } while (0)
 
 // ============================================================================
 // BatchBuilder Tests
