+++ tests/test_json_constrain.cu
@@ -10,12 +10,7 @@
 #include <string>
 #include <cfloat>
 
-#define SKIP_IF_NO_CUDA()                       \
-    do {                                        \
-        int dev;                                \
-        if (cudaGetDevice(&dev) != cudaSuccess) \
-            GTEST_SKIP();                       \
-    } while (0)
+#include "test_cuda_skip.h"
 
 namespace imp {
 namespace {
