+++ tests/test_ssm.cu
@@ -7,12 +7,7 @@
 #include <vector>
 #include <cmath>
 
-#define SKIP_IF_NO_CUDA()                       \
-    do {                                        \
-        int dev;                                \
-        if (cudaGetDevice(&dev) != cudaSuccess) \
-            GTEST_SKIP();                       \
-    } while (0)
+#include "test_cuda_skip.h"
 
 namespace imp {
 namespace {
